requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
openpyxl==3.1.5
SQLAlchemy==2.0.36
python-dotenv==1.0.1
mysqlclient==2.2.5
Werkzeug==3.0.6